requirements.txt
fastapi
uvicorn
gnews
newspaper3k
google-generativeai
requests